+++ part_data.hpp
@@ -14,7 +14,7 @@ class part_grid
          * NumPart should be the number of particles per side; the cube root of the total number of particles.
          * Masses are the mass ratios of the different particle species, for making sure no two types share the same position
          * Box is the boxsize.*/
-        part_grid(const int NumPart[], const double Masses[], const double Box);
+        part_grid(const int64_t NumPart[], const double Masses[], const double Box);
         /* Get the position of a particle at some index on a grid.
          * We want to return:
          * size_t index = k + j* NumPart + i * NumPart**2
@@ -25,13 +25,13 @@ class part_grid
          */
         double Pos(size_t index, int axis, int type);
         //Get number of particles
-        int GetNumPart(int type);
+        int64_t GetNumPart(int type);
         //Get box size
         inline double GetBox(){
             return Box;
         }
     private:
-        const std::valarray<int> NumPart;
+        const std::valarray<int64_t> NumPart;
         const double Box;
         double pspace[N_TYPES];
         double shift[N_TYPES][3];
